Added abstract classes and refactored packaging
